Intake Form Questions Update (#50)
* combined standard questions and outreach, and renamed standard to visit per ahope excel doc; also updated appropriate sliders

* addressing alex's comments


combined standard questions and outreach
renamed standard to visit per ahope excel doc
updated appropriate sliders (number of people helping and age of first injection, though these should later be changed to a different form of input)
